auth, lecture の整理
