+++ src/auth/AuthContext.js
@@ -1,7 +1,7 @@
 import { createContext, useState, useContext, useEffect } from 'react';
-import { auth } from './firebase';
-
+import { auth } from '../firebase';
 const AuthContext = createContext();
+
 
 export function useAuthContext() {
   return useContext(AuthContext);
