+++ src/components/Footer.js
@@ -8,19 +8,19 @@ export function Footer() {
                 <div className="foot">関連団体</div>
                 <div className="link_imgs">
                     <div className='link_img'>
-                        <a rel="noreferrer" href="https://mnes.life/" target="_blank"><img class="img" border="0" width="250" height="125" src={mnesimg} alt="MNES"></img></a>
+                        <a rel="noreferrer" href="https://mnes.life/" target="_blank"><img className="img" border="0" width="250" height="125" src={mnesimg} alt="MNES"></img></a>
                     </div>
                     <div className='link_img'>
-                        <a rel="noreferrer" href="https://laime-ml.github.io/?fbclid=IwAR2ETxkW4ZUq9oRqd7Mn04ffzviU7GYSrS3Ho3SzAsgkB5wmxdrhR8QzLi4" target="_blank"><img class="img" border="1" width="330" height="125" src={laimeimg} alt="LAIME"></img></a>
+                        <a rel="noreferrer" href="https://laime-ml.github.io/?fbclid=IwAR2ETxkW4ZUq9oRqd7Mn04ffzviU7GYSrS3Ho3SzAsgkB5wmxdrhR8QzLi4" target="_blank"><img className="img" border="1" width="330" height="125" src={laimeimg} alt="LAIME"></img></a>
                     </div>
                     <div className='link_img'>
-                        <a rel="noreferrer" href="https://sites.google.com/mnes.org/mnist-official/home" target="_blank"><img class="img" width="240" height="125" src={mnistimg} alt="MNiST"></img></a>
+                        <a rel="noreferrer" href="https://sites.google.com/mnes.org/mnist-official/home" target="_blank"><img className="img" width="240" height="125" src={mnistimg} alt="MNiST"></img></a>
                     </div>
                 </div>
-                <div class="button">
-                    <button class="btn">CONTACT</button>
+                <div className="button">
+                    <button className="btn">CONTACT</button>
                 </div>
-                <div class="bottom">©MNiST</div>
+                <div className="bottom">©MNiST</div>
             </div>
         </footer>
     )
