+++ src/components/lecture/Diagnostic_imaging.js
