+++ src/components/lecture/ExamplePageBasic.js
