+++ src/components/parts/LinkButton.js
@@ -1,4 +1,3 @@
-
 import { Link } from 'react-router-dom';
 
 /**
